+++ tests/testbed/021-dead-candidates/app.rb
@@ -1,0 +1,16 @@
+class Widget
+  validate :check_it
+
+  def check_it
+  end
+
+  def used_once
+  end
+
+  def never_used
+  end
+
+  def run
+    used_once
+  end
+end